Use the update job to resubmit failed products
Instead of using a new job
