+++ src/Product/ProductSyncer.php
@@ -96,8 +96,9 @@ class ProductSyncer implements Service, OptionsAwareInterface {
 			}
 		}
 
-		if ( ! empty( $invalid_products ) ) {
-			do_action( 'gla_batch_update_products_errors', [ $invalid_products ] );
+		$internal_error_products = $this->batch_helper->get_internal_error_products( $invalid_products );
+		if ( ! empty( $internal_error_products ) ) {
+			do_action( 'gla_batch_retry_update_products', [ $internal_error_products ] );
 		}
 
 		do_action(
@@ -164,8 +165,9 @@ class ProductSyncer implements Service, OptionsAwareInterface {
 			}
 		}
 
-		if ( ! empty( $invalid_products ) ) {
-			do_action( 'gla_batch_delete_products_errors', [ $invalid_products ] );
+		$internal_error_products = $this->batch_helper->get_internal_error_products( $invalid_products );
+		if ( ! empty( $internal_error_products ) ) {
+			do_action( 'gla_batch_retry_delete_products', [ $internal_error_products ] );
 		}
 
 		do_action(
